Rename cixtor/golangcilint as alecthomas/golangcilint
I spoke with @alecthomas about the possibility to merge both projects
since his version is older than mine, and people are probably already
using it. My version contains better error and warning parsing, and
works better with the latest version of golangci-lint. He agree to
review and maybe accept a pull-request with my version of the code, so
I’m changing this pull-request to link to his repository rather than
mine.